metadata.h
Comentarios en los metadatos

+++ Práctica 2/dssoo_fs/include/metadata.h
@@ -13,16 +13,16 @@ static inline void bitmap_setbit(char *bitmap_, int i_, int val_) { /*El código
   else
     bitmap_[(i_ >> 3)] &= ~(1 << (i_ & 0x07));
 } 
-
+/* Estructura para el superbloque. Ocupa 5208 bits, 651 bytes. */
 typedef struct {
-	unsigned short numBloquesDatos; /* Número de bloques de datos en el disp. /*Lo ponemos como short porque no hace falta que sea un ing. Ocupa 5 * 2^10*/
+	unsigned short numBloquesDatos; /* Número de bloques de datos en el disp. Lo ponemos como short porque no hace falta que sea un ing. Ocupa 5 * 2^10*/
 	char bmapai[5];
 	char bmapab[640]; 
-	unsigned int tamDispositivo; /* Tamaño total del disp. (en bytes) */ /*Numero de bloques * 2048 (tamaño de bloque) */
+	unsigned int tamDispositivo; /* Tamaño total del disp. (en bytes) Numero de bloques * 2048 (tamaño de bloque) */
 } superbloque;
-
+/* Estructura para cada inodo. Ocupa 304 bits, 38 bytes. */
 typedef struct {
 	char nombre[32]; /* Nombre del fichero/directorio asociado cuya longitud máxima es de 32 caracteres */
 	unsigned int tamanyo; /* Tamaño actual del fichero en bytes */
-	unsigned int bloqueIndirecto; /* array de punteros*/
+	unsigned short bloqueIndirecto; /* Indica el bloque en el que se encuentra el índice de bloques de datos que conforman el fichero*/
 } inodo;
